pyproject.toml
[tool.poetry]
name = "aicup2023-computer-vision"
version = "0.1.0"
description = ""
authors = ["peterxcli <peterxcli@gmail.com>"]
readme = "README.md"
packages = [{include = "aicup2023_computer_vision"}]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.7.0.72"
pandas = "^2.0.1"
pyyaml = "^6.0"
sklearn = "^0.0.post5"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
tensorboard = "^2.13.0"

[tool.poetry.group.dev.dependencies]
torch = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
